requirements.txt
torch
torchvision
pillow
numpy
typing
scikit-image
imageio
huggingface_hub
transformers>=4.39.1
timm
kornia

